+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/ExcessiveUtilities.kt
@@ -6,6 +6,7 @@ import dev.aaronhowser.mods.excessive_utilities.registry.ModRegistries
 import net.minecraft.resources.ResourceLocation
 import net.neoforged.api.distmarker.Dist
 import net.neoforged.fml.ModContainer
+import net.neoforged.fml.ModList
 import net.neoforged.fml.common.Mod
 import net.neoforged.fml.config.ModConfig
 import net.neoforged.neoforge.client.gui.ConfigurationScreen
@@ -20,15 +21,6 @@ class ExcessiveUtilities(
 	modContainer: ModContainer
 ) {
 
-	companion object {
-		const val MOD_ID = "excessive_utilities"
-
-		@JvmField
-		val LOGGER: Logger = LogManager.getLogger(MOD_ID)
-
-		fun modResource(path: String): ResourceLocation = ResourceLocation.fromNamespaceAndPath(MOD_ID, path)
-	}
-
 	init {
 		ModRegistries.register(MOD_BUS)
 
@@ -39,6 +31,19 @@ class ExcessiveUtilities(
 
 		modContainer.registerConfig(ModConfig.Type.SERVER, ServerConfig.CONFIG_SPEC)
 		modContainer.registerConfig(ModConfig.Type.CLIENT, ClientConfig.CONFIG_SPEC)
+	}
+
+	companion object {
+		const val MOD_ID = "excessive_utilities"
+
+		@JvmField
+		val LOGGER: Logger = LogManager.getLogger(MOD_ID)
+
+		fun modResource(path: String): ResourceLocation = ResourceLocation.fromNamespaceAndPath(MOD_ID, path)
+
+		val IS_ATHENA_LOADED: Boolean by lazy {
+			ModList.get().isLoaded("athena")
+		}
 	}
 
 }