+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/datagen/language/ModMenuLang.kt
@@ -41,11 +41,11 @@ object ModMenuLang {
 	const val IGNORE_DAMAGE_OFF = "tooltip.excessive_utilities.item_filter.ignore_damage_off"
 	const val IGNORE_ALL_COMPONENTS_ON = "tooltip.excessive_utilities.item_filter.ignore_all_components_on"
 	const val IGNORE_ALL_COMPONENTS_OFF = "tooltip.excessive_utilities.item_filter.ignore_all_components_off"
-
 	const val SNOW_GLOBE_TOOLTIP_1 = "tooltip.excessive_utilities.magical_snow_globe_1"
 	const val SNOW_GLOBE_TOOLTIP_2 = "tooltip.excessive_utilities.magical_snow_globe_2"
 	const val SNOW_GLOBE_INSTRUCTIONS = "tooltip.excessive_utilities.magical_snow_globe_instructions"
 	const val SNOW_GLOBE_READY = "tooltip.excessive_utilities.magical_snow_globe_ready"
+	const val ATHENA_COMPATIBLE = "tooltip.excessive_utilities.athena_compatible"
 
 	fun add(provider: ModLanguageProvider) {
 		provider.apply {
@@ -108,6 +108,8 @@ object ModMenuLang {
 			add(SNOW_GLOBE_TOOLTIP_2, "Perhaps our universe is not so different?")
 			add(SNOW_GLOBE_INSTRUCTIONS, "To activate this globe, take it to any %d of the following biome types:")
 			add(SNOW_GLOBE_READY, "The Snow Globe has gathered enough information about our world.")
+
+			add(ATHENA_COMPATIBLE, "Install Athena for connected textures!")
 		}
 	}